+++ front-end/src/hooks/index.ts
@@ -3,7 +3,7 @@ export { useAuth } from "./useAuth/useAuth";
 export { default as useBoolean } from "./useBoolean/useBoolean";
 export { default as useBreakpoints } from "./useBreakpoints/useBreakpoints";
 export { default as useFetch } from "./useFetch/useFetch";
-export { default as useForm, Values, SubmitAction } from "./useForm/useForm";
+export { default as useForm, getFormProps, getInputProps, getRadioProps } from "./useForm/useForm";
 export * from "./useHistory/useHistory";
 export { default as useLabId } from "./useLabId/useLabId";
 export { default as useToggle } from "./useToggle/useToggle";
